refactor: break up into multiple files
